// src/app/Models/FirmProposal/partners-info.ts
export class PartnersInfo {
    ID: number;
    CLIENT_ID: number;
    FIRM_INFORMATION_ID: number;
    NAME: string;
    PAN_NUMBER: string;
    AADHAR: string;
    DIRECTORS_IDENTITY_NUMBER: string;
    MOBILE_NUMBER: string;
    GENDER: string = 'M';
    DOB: any;
    CASTE: string;
    RELIGION: string;
    EDUCATION: string;
    MEMBERSHIP_NUMBER: string;
    RELATIONSHIP_WITH_OTHER: string;
    IS_AUTHORISED_TO_TRANSACT: boolean;
}